config site from notion
