--- example/src/components/nav-list.js
@@ -1,121 +1,0 @@
-import React from 'react'
-import PropTypes from 'prop-types'
-import { withStyles } from '@material-ui/core/styles'
-import List from '@material-ui/core/List'
-import ListItem from '@material-ui/core/ListItem'
-import ListItemIcon from '@material-ui/core/ListItemIcon'
-import ListItemText from '@material-ui/core/ListItemText'
-import HomeIcon from '@material-ui/icons/Home'
-import TagIcon from '@material-ui/icons/Bookmark'
-import AboutIcon from '@material-ui/icons/Person'
-import BookListIcon from '@material-ui/icons/LibraryBooks'
-import ArchiveIcon from '@material-ui/icons/Archive'
-import FriendIcon from '@material-ui/icons/People'
-
-import CardMedia from '@material-ui/core/CardMedia'
-import { Link } from 'gatsby'
-
-import config from '../../config'
-
-
-function ListItemLink(props) {
-    return <ListItem button component="a" {...props} />
-}
-
-const styles = theme => ({
-    root: {
-        width: '100%',
-        maxWidth: 360,
-        backgroundColor: theme.palette.background.paper,
-    },
-    media: {
-        height: 200,
-    },
-})
-
-const MyLink = props => <Link {...props} />
-
-function SimpleList(props) {
-    let avatar = require(`../static/${config.blogMeta.avatar}`)
-    const { classes } = props
-
-    return (
-        <div className={classes.root}>
-            <MyLink to={'/'}>
-                <CardMedia
-                    className={classes.media}
-                    image={avatar}
-                    title="Contemplative Reptile"
-                />
-            </MyLink>
-            <List component="nav">
-                <MyLink to={'/'}>
-                    <ListItem button>
-                        <ListItemIcon>
-                            <HomeIcon color={'primary'} />
-                        </ListItemIcon>
-                        <ListItemText primary="主页" />
-                    </ListItem>
-                </MyLink>
-
-                <MyLink to={'tags'}>
-                    <ListItem button>
-                        <ListItemIcon>
-                            <TagIcon style={{ color: '#673ab7' }} />
-                        </ListItemIcon>
-                        <ListItemText primary="标签" />
-                    </ListItem>
-                </MyLink>
-
-                <MyLink to={'archives'}>
-                    <ListItem button>
-                        <ListItemIcon>
-                            <ArchiveIcon style={{ color: '#f44336' }} />
-                        </ListItemIcon>
-                        <ListItemText primary="归档" />
-                    </ListItem>
-                </MyLink>
-
-                <MyLink to={`posts/${config.blogMeta.aboutPostSlug}/`}>
-                    <ListItem button>
-                        <ListItemIcon>
-                            <AboutIcon style={{ color: '#ffc107' }} />
-                        </ListItemIcon>
-                        <ListItemText primary="关于" />
-                    </ListItem>
-                </MyLink>
-                <MyLink to={'books'}>
-                    <ListItem button>
-                        <ListItemIcon>
-                            <BookListIcon style={{ color: '#607d8b' }} />
-                        </ListItemIcon>
-                        <ListItemText primary="书单" />
-                    </ListItem>
-                </MyLink>
-                <MyLink to={'links'}>
-                    <ListItem button>
-                        <ListItemIcon>
-                            <FriendIcon style={{ color: '#e91e63' }} />
-                        </ListItemIcon>
-                        <ListItemText primary="友链" />
-                    </ListItem>
-                </MyLink>
-                <MyLink to={'moments'}>
-                    <ListItem button>
-                        <ListItemIcon>
-                            <FriendIcon style={{ color: '#e91e63' }} />
-                        </ListItemIcon>
-                        <ListItemText primary="动态" />
-                    </ListItem>
-                </MyLink>
-
-            </List>
-        </div>
-    )
-}
-
-SimpleList.propTypes = {
-    classes: PropTypes.object.isRequired,
-}
-
-export default withStyles(styles)(SimpleList)
