--- example/src/pages/moments.js
@@ -1,32 +1,0 @@
-import React from 'react'
-import PropTypes from 'prop-types'
-import { withStyles } from '@material-ui/core/styles'
-import withRoot from 'gatsby-theme-gine-blog/src/withRoot'
-import Layout from 'gatsby-theme-gine-blog/src/components/layout'
-import { graphql } from 'gatsby'
-
-const styles = theme => ({
-    root: {
-        textAlign: 'center',
-        paddingTop: theme.spacing.unit * 20,
-    },
-    index: {
-        margin: '0 auto',
-    },
-})
-
-class Index extends React.Component {
-    render() {
-        return (
-            <Layout title="动态">
-                这里是测试页面
-            </Layout>
-        )
-    }
-}
-
-Index.propTypes = {
-    classes: PropTypes.object.isRequired,
-}
-
-export default withRoot(withStyles(styles)(Index))
