+++ gatsby-theme-gine-blog/package.json
@@ -21,9 +21,11 @@
     "gatsby-plugin-react-helmet": "^3.0.5",
     "gatsby-plugin-sitemap": "^2.0.4",
     "gatsby-source-notion-database": "^0.4.5",
+    "image-downloader": "^3.5.0",
     "jss": "latest",
     "katex": "^0.10.2",
     "lodash": "^4.17.13",
+    "notabase": "^0.6.8",
     "prismjs": "^1.16.0",
     "prop-types": "latest",
     "puppeteer": "^1.17.0",
