+++ gatsby-theme-gine-blog/src/components/link/LinkCard.js
