--- gatsby-theme-gine-blog/src/createDynamicPage.js
@@ -1,13 +1,0 @@
-
-//  
-const config = require('../config')
-
-export default function createDynamicPage(createPage) {
-
-    if (config.draw.open){
-        createPage({
-
-        })
-    }
-
-}
