+++ gatsby-theme-gine-blog/src/pages/books.js
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { withStyles } from '@material-ui/core/styles'
 import withRoot from '../withRoot'
-import Layout from '../components/layout'
+import Layout from '../components/layout/index'
 import BookItem from '../components/book/bookItem'
 import { graphql } from 'gatsby'
