+++ gatsby-theme-gine-blog/src/pages/tags.js
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { withStyles } from '@material-ui/core/styles'
 import withRoot from '../withRoot'
-import Layout from '../components/layout'
+import Layout from '../components/layout/index'
 import countBy from 'lodash/countBy'
 import ColorfulTag, { getHashColor } from '../components/utils/hash-colorful-tag'
 import { graphql } from 'gatsby'
