+++ gatsby-theme-gine-blog/src/static/avatar.jpg
