+++ gatsby-theme-gine-blog/src/withRoot.js
@@ -1,8 +1,8 @@
-import React from 'react';
-import { MuiThemeProvider } from '@material-ui/core/styles';
-import CssBaseline from '@material-ui/core/CssBaseline';
-import JssProvider from 'react-jss/lib/JssProvider';
-import getPageContext from './getPageContext';
+import React from 'react'
+import { MuiThemeProvider } from '@material-ui/core/styles'
+import CssBaseline from '@material-ui/core/CssBaseline'
+import JssProvider from 'react-jss/lib/JssProvider'
+import getPageContext from './getPageContext'
 
 function withRoot(Component) {
   class WithRoot extends React.Component {
